pyproject.toml
# path: pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coupon-causal"
version = "0.1.0"
description = "Production-quality causal impact analysis for retail coupon campaigns"
authors = [{name = "Causal Impact Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas==2.1.4",
    "numpy==1.26.2",
    "scikit-learn==1.3.2",
    "statsmodels==0.14.1",
    "matplotlib==3.8.2",
    "seaborn==0.13.0",
    "plotly==5.18.0",
    "pyyaml==6.0.1",
    "scipy==1.11.4",
    "joblib==1.3.2",
    "tqdm==4.66.1",
    "econml==0.15.0",
    "dowhy==0.11.1",
    "streamlit==1.29.0",
    "pyarrow==14.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
    "jupyter==1.0.0",
    "ipykernel==6.27.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "N806",  # Variable should be lowercase (X, T, Y are standard in causal inference)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
